requirements.txt
aniso8601==9.0.1
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confluent-kafka==2.1.0
Flask==2.2.3
Flask-RESTful==0.3.9
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
protobuf==3.20.3
pydentic==0.0.1.dev3
pyshorteners==1.0.1
python-stdnum==1.18
pytz==2023.3
requests==2.29.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
